Minor rename of internal default values for grouping min/max aggregat… (ESQL-315)
Minor rename of internal default values for grouping min/max aggregators.
Fix issue with negative values in max aggregator.

+++ server/src/main/java/org/elasticsearch/compute/aggregation/DoubleState.java
@@ -26,7 +26,7 @@ final class DoubleState implements AggregatorState<DoubleState> {
         this(0);
     }
 
-    DoubleState(long value) {
+    DoubleState(double value) {
         this.doubleValue = value;
         this.serializer = new DoubleStateSerializer();
     }
